// frontend/src/app/services/openapi/model/invoice-response.ts
/**
 * OpenAPI definition
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: v0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export interface InvoiceResponse { 
    id?: number;
    invoiceType?: InvoiceResponse.InvoiceTypeEnum;
    dateTime?: string;
}
export namespace InvoiceResponse {
    export type InvoiceTypeEnum = 'PURCHASE' | 'CANCELLATION';
    export const InvoiceTypeEnum = {
        Purchase: 'PURCHASE' as InvoiceTypeEnum,
        Cancellation: 'CANCELLATION' as InvoiceTypeEnum
    };
}